# src/frontend/sqeez/src/api/api.yaml
openapi: 3.1.1
info:
  title: Sqeez.Api | v1
  version: 1.0.0
servers:
  - url: http://localhost:5000/
paths:
...
  /api/quiz-attempts/start:
    post:
      tags:
        - QuizAttempts
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartQuizAttemptDto"
          text/json:
            schema:
              $ref: "#/components/schemas/StartQuizAttemptDto"
          application/*+json:
            schema:
              $ref: "#/components/schemas/StartQuizAttemptDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/QuizAttemptDto"
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptDto"
            text/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptDto"
  /api/quiz-attempts/{id}/answer:
    post:
      tags:
        - QuizAttempts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitQuestionResponseDto"
          text/json:
            schema:
              $ref: "#/components/schemas/SubmitQuestionResponseDto"
          application/*+json:
            schema:
              $ref: "#/components/schemas/SubmitQuestionResponseDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/QuestionAnsweredDto"
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionAnsweredDto"
            text/json:
              schema:
                $ref: "#/components/schemas/QuestionAnsweredDto"
  /api/quiz-attempts/{id}/next-question:
    get:
      tags:
        - QuizAttempts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                pattern: ^-?(?:0|[1-9]\d*)$
                type:
                  - "null"
                  - integer
                  - string
                format: int64
            application/json:
              schema:
                pattern: ^-?(?:0|[1-9]\d*)$
                type:
                  - "null"
                  - integer
                  - string
                format: int64
            text/json:
              schema:
                pattern: ^-?(?:0|[1-9]\d*)$
                type:
                  - "null"
                  - integer
                  - string
                format: int64
  /api/quiz-attempts/{id}/complete:
    post:
      tags:
        - QuizAttempts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/QuizAttemptDto"
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptDto"
            text/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptDto"
  /api/quiz-attempts/{id}:
    get:
      tags:
        - QuizAttempts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/QuizAttemptDetailDto"
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptDetailDto"
            text/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptDetailDto"
    delete:
      tags:
        - QuizAttempts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: boolean
            application/json:
              schema:
                type: boolean
            text/json:
              schema:
                type: boolean
  /api/quiz-attempts/quiz/{quizId}:
    get:
      tags:
        - QuizAttempts
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
        - name: pageNumber
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
            default: 20
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/PagedResponseOfQuizAttemptDto"
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponseOfQuizAttemptDto"
            text/json:
              schema:
                $ref: "#/components/schemas/PagedResponseOfQuizAttemptDto"
  /api/quiz-attempts/responses/{responseId}/grade:
    patch:
      tags:
        - QuizAttempts
      parameters:
        - name: responseId
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeQuestionResponseDto"
          text/json:
            schema:
              $ref: "#/components/schemas/GradeQuestionResponseDto"
          application/*+json:
            schema:
              $ref: "#/components/schemas/GradeQuestionResponseDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/QuestionResponseDto"
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionResponseDto"
            text/json:
              schema:
                $ref: "#/components/schemas/QuestionResponseDto"
...
  /api/teachers:
    get:
      tags:
        - Teachers
      parameters:
        - name: Department
          in: query
          schema:
            type: string
        - name: SearchTerm
          in: query
          schema:
            type: string
        - name: IsOnline
          in: query
          schema:
            type: boolean
        - name: SchoolClassId
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
        - name: IsArchived
          in: query
          schema:
            type: boolean
        - name: StrictRoleOnly
          in: query
          schema:
            type: boolean
        - name: PageNumber
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
        - name: PageSize
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/PagedResponseOfTeacherDto"
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponseOfTeacherDto"
            text/json:
              schema:
                $ref: "#/components/schemas/PagedResponseOfTeacherDto"
  /api/teachers/{id}:
    get:
      tags:
        - Teachers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/TeacherDto"
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherDto"
            text/json:
              schema:
                $ref: "#/components/schemas/TeacherDto"
    patch:
      tags:
        - Teachers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchTeacherDto"
          text/json:
            schema:
              $ref: "#/components/schemas/PatchTeacherDto"
          application/*+json:
            schema:
              $ref: "#/components/schemas/PatchTeacherDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/TeacherDto"
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherDto"
            text/json:
              schema:
                $ref: "#/components/schemas/TeacherDto"
    delete:
      tags:
        - Teachers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: boolean
            application/json:
              schema:
                type: boolean
            text/json:
              schema:
                type: boolean
...
components:
  schemas:
...
    AttemptStatus:
      enum:
        - Created
        - Started
        - Completed
        - Abandoned
...
    GradeQuestionResponseDto:
      required:
        - score
        - isLiked
      type: object
      properties:
        score:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        isLiked:
          type: boolean
...
    PagedResponseOfQuizAttemptDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/QuizAttemptDto"
        totalCount:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        pageNumber:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        pageSize:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        totalPages:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
...
    PagedResponseOfTeacherDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TeacherDto"
        totalCount:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        pageNumber:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        pageSize:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        totalPages:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
...
    PatchTeacherDto:
      type: object
      properties:
        department:
          type:
            - "null"
            - string
        managedClassId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int64
        username:
          type:
            - "null"
            - string
        email:
          type:
            - "null"
            - string
        schoolClassId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int64
        avatarUrl:
          type:
            - "null"
            - string
...
    QuestionAnsweredDto:
      required:
        - id
        - quizQuestionId
        - responseTimeMs
        - freeTextAnswer
        - isLiked
        - score
        - selectedOptionIds
      type: object
      properties:
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        quizQuestionId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        responseTimeMs:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        freeTextAnswer:
          type:
            - "null"
            - string
        isLiked:
          type: boolean
        score:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        selectedOptionIds:
          type: array
          items:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
        correctOptionIds:
          type:
            - "null"
            - array
          items:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
        correctFreeTextAnswer:
          type:
            - "null"
            - string
        nextQuestionId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int64
    QuestionResponseDto:
      required:
        - id
        - quizQuestionId
        - responseTimeMs
        - freeTextAnswer
        - isLiked
        - score
        - selectedOptionIds
      type: object
      properties:
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        quizQuestionId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        responseTimeMs:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        freeTextAnswer:
          type:
            - "null"
            - string
        isLiked:
          type: boolean
        score:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        selectedOptionIds:
          type: array
          items:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
    QuizAttemptDetailDto:
      required:
        - id
        - quizId
        - enrollmentId
        - startTime
        - endTime
        - status
        - totalScore
        - mark
        - responses
      type: object
      properties:
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        quizId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        enrollmentId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        startTime:
          type:
            - "null"
            - string
          format: date-time
        endTime:
          type:
            - "null"
            - string
          format: date-time
        status:
          $ref: "#/components/schemas/AttemptStatus"
        totalScore:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        mark:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int32
        responses:
          type: array
          items:
            $ref: "#/components/schemas/QuestionResponseDto"
    QuizAttemptDto:
      required:
        - id
        - quizId
        - enrollmentId
        - startTime
        - endTime
        - status
        - totalScore
        - mark
      type: object
      properties:
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        quizId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        enrollmentId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        startTime:
          type:
            - "null"
            - string
          format: date-time
        endTime:
          type:
            - "null"
            - string
          format: date-time
        status:
          $ref: "#/components/schemas/AttemptStatus"
        totalScore:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        mark:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int32
        nextQuestionId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int64
...
    StartQuizAttemptDto:
      required:
        - quizId
        - enrollmentId
      type: object
      properties:
        quizId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        enrollmentId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
...
    SubmitQuestionResponseDto:
      required:
        - quizQuestionId
        - responseTimeMs
        - freeTextAnswer
        - selectedOptionIds
      type: object
      properties:
        quizQuestionId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        responseTimeMs:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        freeTextAnswer:
          type:
            - "null"
            - string
        selectedOptionIds:
          type: array
          items:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
...
    TeacherDto:
      type: object
      properties:
        department:
          type:
            - "null"
            - string
        managedClassId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int64
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        username:
          type: string
        email:
          type: string
        currentXP:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        role:
          type: string
        avatarUrl:
          type:
            - "null"
            - string
        lastSeen:
          type: string
          format: date-time
        schoolClassId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int64
...